Revert "AddKoreAttribute as a Frontend Kompiler Pass (#3522)"
This reverts commit 4db2ccbc84c88be94f4ac3b78849c147418d997c.
