+++ kernel/src/main/java/org/kframework/backend/kore/KoreBackend.java
@@ -99,7 +99,6 @@ public class KoreBackend extends AbstractBackend {
           mainModule = ModuleTransformer.fromSentenceTransformer(new MinimizeTermConstruction(mainModule)::resolve, "Minimize term construction").apply(mainModule);
         }
         mainModule = ModuleTransformer.from(new GenerateMapCeilAxioms(mainModule, kompileOptions)::gen, "Generate map ceil axioms").apply(mainModule);
-        mainModule = ModuleTransformer.fromSentenceTransformerAtt(new AddKoreAttributes(mainModule, kompileOptions)::add, "Add kore attributes").apply(mainModule);
 
         return mainModule;
     }
